merge into lemon

+++ src/main/java/com/lemon/supershop/swp391fa25evdm/category/model/dto/CategoryReq.java
@@ -9,7 +9,7 @@ public class CategoryReq {
     private String brand;
     private String version;
     private String type;
-    private Double basePrice;
+    private Long basePrice;
     private Integer warranty;
     private Boolean isSpecial;
     private String description;
@@ -50,11 +50,11 @@ public class CategoryReq {
         this.type = type;
     }
 
-    public Double getBasePrice() {
+    public Long getBasePrice() {
         return basePrice;
     }
 
-    public void setBasePrice(Double basePrice) {
+    public void setBasePrice(Long basePrice) {
         this.basePrice = basePrice;
     }
